requirements.txt
requests~=2.27.1
discord.py==1.7.3
d20==1.1.2